docker-compose-onboarding/app/requirements.txt
mysql-connector-python
pandas
boto3
metaflow